Calculate node GUIDs using Outputs if available
When calculating the GUID for a node, generate it by hashing the Outputs, if available. This ensures that as a node's inputs or action are changed over successive build runs, we can still identify the node consistently, and retrieve the data about its inputs from the previous run in order to report the differences.

If the node has no Outputs, then fall back to using the previous behaviour.