Record node indices and action root PIDs in profiler output
When profiling is enabled, capture both the DAG node index that a profiler entry corresponds to (if appropriate), and for action nodes that launch external processes, capture the PID of the process that has been launched.

The PID will allow us to correlate profiler data with system tracing tools like Process Monitor or DTrace, and the node index will let us connect it back to the DAG, giving us a pipeline for connecting DAG nodes to their resulting process and IO activity. We can then explore using this to verify that a node's inputs and outputs are fully described.