Add profiling for CheckInputSignature
There's a big block of unaccounted-for build time in zero-change builds for large DAGs. Code inspection, and profiling with the VS profiler, revealed that it's primarily the call tree under CheckInputSignature - stat64() calls for most files, plus some file hashing. By adding a profiler scope to CheckInputSignature, we account for most of the unaccounted-for time.