Shared Resources support
Support for Shared Resources, which are logical entities which can be used by multiple build actions (e.g. compiler servers). A resource can run one command when it is created, and another when it is shut down. The first time we try to execute an action that needs a particular resource, the resource is created; and at the end of the build, all created resources are destroyed.