Stable ordering + dedup for implicit inputs
When collecting implicit inputs for a node's input signature, accumulate them in a hashset and then add them to the signature after scanning all explicit input files, rather than doing it on-the-fly. This allows us to ensure a stable ordering on the implicit inputs, as well as deduplicating them.

Fix the way implicit inputs are recorded into the buildstate to also be deduplicated and stably ordered, to match.