ExecWin32: on temp file creation error, abort sooner rather than later
Previously, the INVALID_HANDLE_VALUE would be passed to CreateProcessW,
which would fail with ERROR_INVALID_PARAMETER, and then cause Tundra to
abort. There's no reason to not simply fail immediately.

Also, use CroakErrno to include the actual failure reason with the error
message. (In my case, it was ERROR_SHARING_VIOLATION because of the file
being kept open by a process lingering from a previous build action.)