Filter the files used to calculate a glob signature
It's common that the files that determine whether the frontend must be rerun are not actually 'all files,' but instead 'all files that match a particular pattern,' e.g. "*.bee.cs". This change allows the frontend to optionally provide a filter expression for each GlobSignature which will be used to filter the files that should be accounted for.